# Minor improvements for Target<T>