--- a/Ninject.sln
+++ b/Ninject.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26730.16
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.28315.86
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{3E299B94-5F07-49DE-8226-55EDC56F13E3}"
 	ProjectSection(SolutionItems) = preProject
@@ -17,6 +17,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestAssembly", "src\TestAssembly\TestAssembly.csproj", "{339DD924-374A-4221-8671-BE951AD280F9}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestModules", "src\TestModules\TestModules.csproj", "{3D6710B4-726A-44B2-A4A2-CCF3882FC611}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ninject.Benchmarks", "src\Ninject.Benchmarks\Ninject.Benchmarks.csproj", "{02241393-6742-44AA-904A-095E654E8D0A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -40,6 +42,10 @@ Global
 		{3D6710B4-726A-44B2-A4A2-CCF3882FC611}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3D6710B4-726A-44B2-A4A2-CCF3882FC611}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3D6710B4-726A-44B2-A4A2-CCF3882FC611}.Release|Any CPU.Build.0 = Release|Any CPU
+		{02241393-6742-44AA-904A-095E654E8D0A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{02241393-6742-44AA-904A-095E654E8D0A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{02241393-6742-44AA-904A-095E654E8D0A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{02241393-6742-44AA-904A-095E654E8D0A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE